restgateway/requirements.txt
Flask
connexion
swagger_ui_bundle
connexion[swagger-ui]
Flask-Injector
fastavro
elasticsearch
boto
grpcio
grpcio-tools
numpy
